Make BBTree.setLocation faster.
Each BBTree was tracking its own copy of rootX and rootY; setLocation
would recurse through the tree, setting each BBTree instance's copies.
This was slow, and probably wasted a chunk of memory, too. Instead, give
each BBTree a reference to its parent, so we can recurse up the tree,
and get the root BBTree's rootX & rootY. This is probably slower when we
read the rootX & rootY, but faster when we setLocation (which we do a lot).

I originally tried giving each BBTree a reference directly to its root,
but because BBTreeBuilder builds bottom-up, the parent BBTree doesn't have
its root, at the point where it needs to pass it to its kids. There's maybe
some better way to do that, but this is pretty good for now.