Adding a writeToSVG Method to WordCram.
The method can be called after the drawing has happened.
The applet (or svg) will still show the cram.
The reason the graphics is not substituted is that the streaming into the svg file has to happen afterwards and I do not see a clever way of doing this wihtout user interaction.

The upside is that the user can also see the cram.

The downside is:
- The words get drawn twice. In the test it is fast as they do not need to get placed, but still
- I had to make the eWords available to WordCram