WordShaper: sensibly default the Font Size parameter.
If you're passing a Font, you probably know what you're doing,
so use the font sized as it is.

If you're passing a PFont, default it to the size it's already
at. (You're probably shaping a big ol' word for a
ShapeBasedImage.)

But, if you're passing a PFont, allow a font-size override.
This is for the WordCramEngine, which has to size the font
for each word, based on the WordSizer.